[FEATURE] Add build-env.sh script
With this development flow, the user is required to change their IP
address in the env file so that the docker workflow will run.  This is
because the database lives on the host, not inside another docker file
which means the DATABASE_URL needs an IP address instead of loalhost.
This script makes it easy for a developer to quickly update their .env
file so they do not have to manually go through this.